% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageYears.R
\name{age_years}
\alias{age_years}
\title{\code{age_years} Age in years from birthdata and second date.}
\usage{
age_years(birthdate, to)
}
\arguments{
\item{birthdate}{Birthdate.}

\item{to}{Date to (procedure date).}
}
\value{
Age in whole years.
}
\description{
\code{age_years} Age in years from birthdata and second date.
}
\examples{
age_years(birthdate = "1960-12-20", to = "2016-02-18")
}

