% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LuiPower.R
\name{fwr}
\alias{fwr}
\title{\code{fwr} power and omega from Lui(2012)}
\usage{
fwr(n, d, ud, ttail = 0.975)
}
\arguments{
\item{n}{Sample Size (can be a range).}

\item{d}{Delta/Sigma.}

\item{ud}{U/Sigma.}

\item{ttail}{2-sided upper tail (defaults to 0.975)}
}
\value{
Power and Omega.
}
\description{
\code{fwr} power and omega from Lui(2012)
}
\examples{
fwr(n = 70, d = 0.5, ud = 0.35, ttail = 0.975)
fwr(n = seq(65, 75, by=1), d = 0.5, ud = 0.35, ttail = 0.975)

}

