% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petemisc.R
\docType{package}
\name{petemisc}
\alias{petemisc}
\alias{petemisc-package}
\title{petemisc: A package of misc functions.}
\description{
petemisc: A package of misc functions.
}
\section{petemisc functions}{

\code{\link{fwr}} returns power and omega from Lui(2012).
\code{\link{age_years}} returns age in years from birthdata and second date.
}

