% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LuiPower.R
\name{fwr}
\alias{fwr}
\title{\code{fwr} power and omega from Lui(2012)}
\usage{
fwr(n, d, ud)
}
\arguments{
\item{n}{Sample Size.}

\item{d}{Delta/Sigma.}

\item{ud}{U/Sigma.}
}
\value{
Power and Omega.
}
\description{
\code{fwr} power and omega from Lui(2012)
}
\examples{
fwr(n = 70, d = 0.5, ud = 0.35)
}

